Make ROLLBAR_TOKEN optional | ENG-7652
Since the 'tinkerbell' suite is no longer solely utilized by Packet/Equinix, it
stands to reason not every user/deployment will wish to integrate with rollbar
for exception logging.    This change makes the rollbar integration optional.

Additionally the example log function tests were moved to their own file.
