+++ log/log.go
@@ -57,10 +57,16 @@ func buildConfig(c zap.Config) (*zap.Logger, error) {
 func configureLogger(l *zap.Logger, service string) (Logger, error) {
 	l = l.With(zap.String("service", service))
 
-	rollbarClean := rollbar.Setup(l.Sugar().With("pkg", "log"), service)
 	cleanup := func() {
-		rollbarClean()
 		_ = l.Sync()
+	}
+
+	if ok := os.Getenv("ROLLBAR_TOKEN"); ok != "" {
+		rollbarClean := rollbar.Setup(l.Sugar().With("pkg", "log"), service)
+		cleanup = func() {
+			rollbarClean()
+			_ = l.Sync()
+		}
 	}
 
 	return Logger{service: service, s: l.Sugar(), cleanup: cleanup}.AddCallerSkip(1), nil
@@ -97,7 +103,9 @@ func (l Logger) Close() {
 // All the values of arg are stringified and concatenated without any strings.
 // If no args are provided err.Error() is used as the log message.
 func (l Logger) Error(err error, args ...interface{}) {
-	rollbar.Notify(err, args)
+	if ok := os.Getenv("ROLLBAR_TOKEN"); ok != "" {
+		rollbar.Notify(err, args)
+	}
 	if len(args) == 0 {
 		args = append(args, err)
 	}
